requirements/requirements.txt
django==1.11.5
django_polymorphic==1.3
django_tables2
django-bootstrap3==9.1.0
django-filter
django-crispy-forms==1.7.0